feat: SDK audit fixes — exception hierarchy, vectorization, Rust CLI, terrain CLI
Python SDK:
- Add OpenZenithError, TileNotFoundError, TileDecodeError, NetworkError,
  DataError exception hierarchy in __init__.py
- Replace 14 bare 'except Exception' swallows across elevation.py, fuse.py,
  tracing.py, backends/ozt2.py with structured logging + specific types
- Vectorize load_elevation_grid pixel loop (50-100x faster for large grids)
- Vectorize tpi(), tri(), roughness() neighbor loops via np.stack (5-10x)
- Vectorize terrain_to_glb per-vertex palette lookup with np.searchsorted
- Add tpi, roughness, curvature CLI commands
- Fix COMP_ZSTD aliasing (COMP_ZSTD_V2 for tile_format_v2)
- Fix undefined 'elev' and 'err' references in elevation.py and cli.py
- Register pytest 'integration' mark; auto-deselect integration tests
- Add ruff --fix auto-fixes (unused imports, etc.)

Rust openzenith-core:
- Fix viewshed bilinear nodata handling: proper weighted interpolation
  instead of averaging valid corners (correctness fix)
- Add stream-order CLI command
- Add gradient-predict CLI command
- Add stream_order() and gradient_predict() Python bindings
- Fix viewshed test to verify hill blocking
- Fix clippy: unnecessary f32 casts in ozt2.rs, SQRT_2 const in d8.rs,
  deprecated into_raw_vec() in main.rs
- All 15 Rust tests pass, clippy clean

+++ openzenith-core/src/ozt2.rs
@@ -117,14 +117,12 @@ pub fn gradient_predict(elevation: &ArrayView2<f32>, nodata: f32) -> Array2<i16>
             let (pred, pred_count) = if i == 0 && j == 0 {
                 (0.0, 0)
             } else if i == 0 {
-                (elevation[[0, j - 1]] as f32, 1)
+                (elevation[[0, j - 1]], 1)
             } else if j == 0 {
-                (elevation[[i - 1, 0]] as f32, 1)
+                (elevation[[i - 1, 0]], 1)
             } else {
                 (
-                    elevation[[i, j - 1]] as f32
-                        + elevation[[i - 1, j]] as f32
-                        - elevation[[i - 1, j - 1]] as f32,
+                    elevation[[i, j - 1]] + elevation[[i - 1, j]] - elevation[[i - 1, j - 1]],
                     3,
                 )
             };
